+++ package.js
@@ -1,7 +1,7 @@
 Package.describe({
 	name: 'konecty:user-presence',
 	summary: 'Track user status',
-	version: '1.2.9',
+	version: '1.3.0',
 	git: 'https://github.com/Konecty/meteor-user-presence'
 });
 
@@ -10,6 +10,7 @@ Package.onUse(function(api) {
 
 	api.use('nooitaf:colors@1.1.2_1');
 	api.use('underscore');
+	api.use('tracker');
 
 	api.addFiles('common/common.js');
 	api.addFiles('server/server.js', ['server']);
@@ -17,7 +18,7 @@ Package.onUse(function(api) {
 	api.addFiles('client/client.js', ['client']);
 
 	api.export(['UserPresence'], ['server', 'client']);
-	api.export(['UserPresenceMonitor'], ['server']);
+	api.export(['UserPresenceMonitor', 'UserPresenceEvents'], ['server']);
 });
 
 Package.onTest(function(api) {