+++ server/monitor.js
@@ -1,4 +1,7 @@
 /* globals UserPresenceMonitor, UsersSessions */
+var EventEmitter = Npm.require('events');
+
+UserPresenceEvents = new EventEmitter();
 
 UserPresenceMonitor = {
 	callbacks: [],
@@ -7,27 +10,22 @@ UserPresenceMonitor = {
 	 * The callback will receive the following parameters: user, status, statusConnection
 	 */
 	onSetUserStatus: function(callback) {
-		this.callbacks.push(callback);
+		UserPresenceEvents.on('setUserStatus', callback);
 	},
 
-	runCallbacks: function(user, status, statusConnection) {
-		this.callbacks.forEach(function(callback) {
-			callback.call(null, user, status, statusConnection);
-		});
-	},
-
-	start: function() {
-		UsersSessions.find({}).observe({
-			added: function(record) {
-				UserPresenceMonitor.processUserSession(record, 'added');
-			},
-			changed: function(record) {
-				UserPresenceMonitor.processUserSession(record, 'changed');
-			},
-			removed: function(record) {
-				UserPresenceMonitor.processUserSession(record, 'removed');
+	getUserStatus: function(connections) {
+		var currentStatus = 'offline';
+		connections.forEach(function(connection) {
+			if (connection.status === 'online') {
+				currentStatus = 'online';
+			} else {
+				if (connection.status === 'away' && currentStatus === 'offline') {
+					currentStatus = 'away';
+				}
 			}
 		});
+
+		return currentStatus;
 	},
 
 	processUserSession: function(record, action) {
@@ -36,11 +34,7 @@ UserPresenceMonitor = {
 		}
 
 		if (record.connections == null || record.connections.length === 0 || action === 'removed') {
-			if (record.visitor === true) {
-				UserPresenceMonitor.setVisitorStatus(record._id, 'offline');
-			} else {
-				UserPresenceMonitor.setUserStatus(record._id, 'offline');
-			}
+			UserPresenceMonitor.setUserStatus(record, 'offline');
 
 			if (action !== 'removed') {
 				UsersSessions.remove({_id: record._id, 'connections.0': {$exists: false} });
@@ -48,20 +42,9 @@ UserPresenceMonitor = {
 			return;
 		}
 
-		var connectionStatus = 'offline';
-		record.connections.forEach(function(connection) {
-			if (connection.status === 'online') {
-				connectionStatus = 'online';
-			} else if (connection.status === 'away' && connectionStatus === 'offline') {
-				connectionStatus = 'away';
-			}
-		});
+		var currentStatus = UserPresenceMonitor.getUserStatus(record.connections);
 
-		if (record.visitor === true) {
-			UserPresenceMonitor.setVisitorStatus(record._id, connectionStatus);
-		} else {
-			UserPresenceMonitor.setUserStatus(record._id, connectionStatus);
-		}
+		UserPresenceMonitor.setUserStatus(record, currentStatus, currentStatus);
 	},
 
 	processUser: function(id, fields) {
@@ -76,37 +59,10 @@ UserPresenceMonitor = {
 		}
 	},
 
-	setUserStatus: function(userId, status) {
-		var user = Meteor.users.findOne(userId),
+	setUserStatus: function(session, status, statusConnection) {
+		if (typeof statusConnection === 'undefined') {
 			statusConnection = status;
-
-		if (!user) {
-			return;
 		}
-
-		if (user.statusDefault != null && status !== 'offline' && user.statusDefault !== 'online') {
-			status = user.statusDefault;
-		}
-
-		var query = {
-			_id: userId,
-			$or: [
-				{status: {$ne: status}},
-				{statusConnection: {$ne: statusConnection}}
-			]
-		};
-
-		var update = {
-			$set: {
-				status: status,
-				statusConnection: statusConnection
-			}
-		};
-
-		Meteor.users.update(query, update);
-
-		this.runCallbacks(user, status, statusConnection);
-	},
-
-	setVisitorStatus: function(/*id, status*/) {}
+		UserPresenceEvents.emit('setStatus', session, status, statusConnection);
+	}
 };